+++ app/responses/bloom_trade_client/get_order_response.rb
@@ -1,0 +1,13 @@
+module BloomTradeClient
+  class GetOrderResponse < BaseResponse
+
+    attribute :order, BloomTradeClient::Order, default: :default_order
+
+    private
+
+    def default_order
+      BloomTradeClient::Order.new(body)
+    end
+
+  end
+end