+++ lib/bloom_trade_client/client.rb
@@ -7,6 +7,7 @@ module BloomTradeClient
 
     api_action :get_quote
     api_action :update_quote
+    api_action :get_order, args: [:memo]
 
     private