+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `#get_order` to fetch order details
+
 ## [1.0.0] - 2019-02-22
 ### Changed
 - `BloomTradeClient::ExchangeRates::Convert` raises an error if rate has